refactor: move paypal payment event listener to directory
