+++ src/Controller/PayPalWebhooks.php
@@ -7,8 +7,8 @@ use Give\PaymentGateways\PayPalCommerce\Repositories\MerchantDetails;
 use Give\PaymentGateways\PayPalCommerce\Repositories\Webhooks;
 use Give\PaymentGateways\PayPalCommerce\Webhooks\Listeners\EventListener;
 use Give\PaymentGateways\PayPalCommerce\Webhooks\Listeners\PaymentCaptureCompleted;
-use Give\PaymentGateways\PayPalCommerce\Webhooks\Listeners\PaymentCaptureDenied;
-use Give\PaymentGateways\PayPalCommerce\Webhooks\Listeners\PaymentCaptureRefunded;
+use Give\PaymentGateways\PayPalCommerce\Webhooks\Listeners\PayPalCommerce\PaymentCaptureDenied;
+use Give\PaymentGateways\PayPalCommerce\Webhooks\Listeners\PayPalCommerce\PaymentCaptureRefunded;
 use InvalidArgumentException;
 
 class PayPalWebhooks {
